+++ w_open.h
@@ -2,6 +2,6 @@
 #define _W_OPEN_H_
 
 int w_open_init(int);
-int w_open_job(int, int, int);
+long w_open_job(int, int, int);
 
 #endif /* !_W_OPEN_H_ */
