+++ w_rename.h
@@ -2,6 +2,6 @@
 #define _W_RENAME_H_
 
 int w_rename_init(int);
-int w_rename_job(int, int, int);
+long w_rename_job(int, int, int);
 
 #endif /* !_W_RENAME_H_ */
